minor changes to sound stuff

+++ sound.h
@@ -8,14 +8,14 @@
 // Structure to hold the offsets for each voice's data
 typedef struct {
     uint32_t offset;
-    uint32_t voice1_offset;
-    int16_t voice1_duration;
-    uint32_t voice2_offset;
-    int16_t voice2_duration;
-    uint32_t voice3_offset;
-    int16_t voice3_duration;
-    uint32_t noise_offset;
-    int16_t noise_duration;
+    uint16_t voice1_offset;
+    uint16_t voice1_duration;
+    uint16_t voice2_offset;
+    uint16_t voice2_duration;
+    uint16_t voice3_offset;
+    uint16_t voice3_duration;
+    uint16_t noise_offset;
+    uint16_t noise_duration;
 } Sound;
 
 // Function to init an AGI sound
